<FIX> Paging 옵션에서 Sort 옵션 null 허용

+++ src/main/java/com/instream/tenant/domain/common/domain/dto/PaginationInfoDto.java
@@ -15,5 +15,5 @@ public class PaginationInfoDto<T extends CollectionDto> extends PaginationDto<T>
 
     @Schema(description = "총 개수, 0부터 시작")
     @Min(value = 0)
-    int totalElementCount;
+    long totalElementCount;
 }
